requirements.txt
streamlit
pandas
plotly
gspread
gspread_dataframe
google-auth
